Upgrade to nom 4.0, include linenum info, adjust API
This is a very large commit. The major change is a move from nom3
to nom4 including nom_locate in the mix. The library no longer
offers a parser per se but a tokenizer, having ambitions to do
parsing over the token stream in a later step. The README is now
updated with work that needs to be done in the tokenizer.

Signed-off-by: Brian L. Troutwine <briant@unity3d.com>